+++ lib/puppet/provider/cisco_command_config/cisco.rb
@@ -85,4 +85,40 @@ Puppet::Type.type(:cisco_command_config).provide(:cisco) do
     info "Successfully updated:\n#{e.previous.join("\n")}" unless e.previous.empty?
     raise
   end # command=
+
+  def test_get
+    # This method is for beaker use only. It allows beaker to retrieve any
+    # configuration it needs from the device using puppet resource. Callers
+    # must pass a filter string to test_get.
+    # Example usage:
+    #  puppet resource cisco_command_config 'cc' test_get='incl feature'
+    cmd = 'show running-config all'
+    cmd << " | #{@resource[:test_get]}" if @resource[:test_get]
+
+    output = @node.get(command: cmd)
+    debug "@node.get output:\n#{output}"
+    "\n" + output unless output.nil?
+  end
+
+  def test_get=(noop)
+    # This is a dummy "setter" for test_get(), which is a get-only property.
+    # This dummy method is necessary to keep Puppet from raising an error or
+    # displaying noise.
+  end
+
+  def test_set
+    # This is a dummy "getter" for test_set=(), which is a set-only property.
+    # This dummy method is necessary to keep Puppet from raising an error or
+    # displaying noise.
+  end
+
+  def test_set=(cmds)
+    # This method is for beaker use only. It allows beaker to set simple raw
+    # configuration using puppet resource.
+    # Example usage:
+    #  puppet resource cisco_command_config 'cc' test_set='no feature foo'
+    return if cmds.empty?
+    output = @node.set(values: cmds)
+    debug "@node.set output:\n#{output}" unless output.nil?
+  end
 end # Puppet::Type