+++ tests/beaker_tests/cisco_fabricpath_topology/test_fabricpath_topology.rb
@@ -159,8 +159,11 @@ def test_harness_fabricpath_topology(tests, id)
 end
 
 def testbed_cleanup(agent)
+  logger.info("\n* testbed cleanup")
+  cmds = ['feature nv overlay', 'feature-set fabricpath']
+  config_find_remove(agent, cmds, 'incl ^feature')
+
   remove_all_vlans(agent)
-  resource_absent_cleanup(agent, 'cisco_fabricpath_topology')
 end
 
 #################################################################